pyproject.toml
[project]
name = "warwickcsnotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "jinja2>=3.1.6",
    "markdown>=3.10.2",
    "pypandoc_binary>=1.13", # import is just "pypandoc" but binary so it install pandoc too
]
